Move preferences test into their own directory
