+++ bookwyrm/tests/views/preferences/test_edit_user.py
@@ -1,12 +1,10 @@
 """ test for app action functionality """
-import json
 import pathlib
 from unittest.mock import patch
 from PIL import Image
 from tidylib import tidy_document
 
 from django.contrib.auth.models import AnonymousUser
-from django.contrib.sessions.middleware import SessionMiddleware
 from django.core.files.base import ContentFile
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django.template.response import TemplateResponse
@@ -122,62 +120,7 @@ class EditUserViews(TestCase):
         )
         image = Image.open(image_file)
 
-        result = views.edit_user.crop_avatar(image)
+        result = views.preferences.edit_user.crop_avatar(image)
         self.assertIsInstance(result, ContentFile)
         image_result = Image.open(result)
         self.assertEqual(image_result.size, (120, 120))
-
-    def test_change_password_page(self, _):
-        """there are so many views, this just makes sure it LOADS"""
-        view = views.ChangePassword.as_view()
-        request = self.factory.get("")
-        request.user = self.local_user
-        result = view(request)
-        self.assertIsInstance(result, TemplateResponse)
-        html = result.render()
-        _, errors = tidy_document(html.content)
-        if errors:
-            raise Exception(errors)
-        self.assertEqual(result.status_code, 200)
-
-    def test_delete_user_page(self, _):
-        """there are so many views, this just makes sure it LOADS"""
-        view = views.DeleteUser.as_view()
-        request = self.factory.get("")
-        request.user = self.local_user
-        result = view(request)
-        self.assertIsInstance(result, TemplateResponse)
-        html = result.render()
-        _, errors = tidy_document(html.content)
-        if errors:
-            raise Exception(errors)
-        self.assertEqual(result.status_code, 200)
-
-    @patch("bookwyrm.suggested_users.rerank_suggestions_task")
-    def test_delete_user(self, *_):
-        """use a form to update a user"""
-        view = views.DeleteUser.as_view()
-        form = forms.DeleteUserForm()
-        form.data["password"] = "password"
-        request = self.factory.post("", form.data)
-        request.user = self.local_user
-        middleware = SessionMiddleware()
-        middleware.process_request(request)
-        request.session.save()
-
-        self.assertIsNone(self.local_user.name)
-        with patch(
-            "bookwyrm.models.activitypub_mixin.broadcast_task.delay"
-        ) as delay_mock:
-            view(request)
-        self.assertEqual(delay_mock.call_count, 1)
-        activity = json.loads(delay_mock.call_args[0][1])
-        self.assertEqual(activity["type"], "Delete")
-        self.assertEqual(activity["actor"], self.local_user.remote_id)
-        self.assertEqual(
-            activity["cc"][0], "https://www.w3.org/ns/activitystreams#Public"
-        )
-
-        self.local_user.refresh_from_db()
-        self.assertFalse(self.local_user.is_active)
-        self.assertEqual(self.local_user.deactivation_reason, "self_deletion")
